Fixes to build
