+++ frontend/src/app/shared/components/grids/widgets/add/add.modal.html
@@ -1,9 +1,5 @@
 <div
   class="spot-modal loading-indicator--location"
-  tabindex="0"
-  cdkFocusInitial
-  cdkTrapFocus
-  [cdkTrapFocusAutoCapture]="true"
   data-indicator-name="modal"
 >
   <div class="spot-modal--header">{{text.title}}</div>
