+++ frontend/src/app/shared/components/modal/modal-overlay.component.ts
@@ -30,20 +30,27 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef, Component, ElementRef, OnInit, ViewChild,
 } from '@angular/core';
+import {
+  ComponentPortal,
+  ComponentType,
+  PortalInjector,
+} from '@angular/cdk/portal';
 import { I18nService } from 'core-app/core/i18n/i18n.service';
 import { OpModalService } from 'core-app/shared/components/modal/modal.service';
 
 @Component({
-  templateUrl: './modal-portal.component.html',
+  templateUrl: './modal-overlay.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class OpModalPortalComponent implements OnInit {
+export class OpModalOverlayComponent implements OnInit {
   // override superclass
   // Allowing outside clicks to close the modal leads to the user involuntarily closing
   // the modal when removing error messages or clicking on labels e.g. in the registration modal.
   public closeOnOutsideClick = false;
 
   @ViewChild('portalOutlet') portalOutlet: ElementRef<HTMLElement>;
+
+  modalPortal$ = this.modalService.activeModal$;
 
   constructor(
     readonly modalService:OpModalService,
@@ -53,7 +60,38 @@ export class OpModalPortalComponent implements OnInit {
   ) { }
 
   ngOnInit():void {
+    // onclose:
+    // this.active.closingEvent.emit(this.active);
+    // onopen:
+    // this.active.onOpen();
+    
+    const portal = new ComponentPortal(modal, null, this.injectorFor(injector, locals));
+
+    setTimeout(() => {
+      // Focus on the first element
+      this.active && this.active.onOpen();
+
+      // Mark that we've opened the modal now
+      this.opening = false;
+
+      // Trigger another round of change detection in the modal
+      ref.changeDetectorRef.detectChanges();
+    }, 20);
   }
 
-  modalPortal$ = this.modalService.activeModal$;
+  /**
+   * Create an augmented injector that is equal to this service's injector + the additional data
+   * passed into +show+.
+   * This allows callers to pass data into the newly created modal.
+   */
+  private injectorFor(injector:Injector, data:Record<string, unknown>) {
+    const injectorTokens = new WeakMap();
+    // Pass the service because otherwise we're getting a cyclic dependency between the portal
+    // host service and the bound portal
+    data.service = this;
+
+    injectorTokens.set(OpModalLocalsToken, data);
+
+    return new PortalInjector(injector, injectorTokens);
+  }
 }
